cleanup AppMode and test
