+++ LuasWatch Watch App/LuasWatchApp.swift
@@ -3,11 +3,10 @@
 //  Copyright © 2023 mApps.ie. All rights reserved.
 //
 
-import SwiftData
-import SwiftUI
-
 import LuasAPI
 import LuasApp
+import SwiftData
+import SwiftUI
 
 @main
 struct LuasWatch_Watch_App: App {
